backend/requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
reportlab>=4.0.0
qrcode>=7.4.0
Pillow>=10.0.0
daphne>=4.0.0
twilio>=9.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
channels>=4.0.0
py-algorand-sdk>=2.6.0
algokit-utils>=3.0.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
